Safeguarded dangerous shell command

+++ releng/org.eclipse.rap.releng/scripts/runtime-hudson-1.2.sh
@@ -6,7 +6,7 @@
 #
 
 # Cleanup workspace dir
-rm -rf "$WORKSPACE"/*
+test -d "$WORKSPACE" && rm -rf "$WORKSPACE"/*
 
 runtimeDir=/home/build/rap/build-runtimes/eclipse-3.4.1
 scriptsDir=`dirname $0`
