+++ releng/org.eclipse.rap.releng/scripts/tooling-hudson-1.2.sh
@@ -6,7 +6,7 @@
 #
 
 # Cleanup workspace dir
-rm -rf "$WORKSPACE"/*
+test -d "$WORKSPACE" && rm -rf "$WORKSPACE"/*
 
 runtimeDir=/home/build/rap/build-runtimes/eclipse-3.4.1
 rapTargets="/home/build/.hudson/jobs/RAP Runtime 1.2 (Galileo)/lastSuccessful/archive"
